recursive heap w/ cancel bid + tests
