+++ programs/caravan/src/state/mod.rs
@@ -1,1 +1,2 @@
 pub mod heap;
+pub mod test;
